refactor(core): 统一布尔字段命名及调整方法参数类型
- 将布尔类型字段统一添加 is 前缀，提升代码可读性
- 将 calculateBackoff 方法的 retryCount 参数类型由 int 改为 long，并在方法内转换类型
- 修正 PersistentDeadLetterQueue 中 dbAvailable 字段名为 isDbAvailable，保持命名一致
- 代码风格和命名规范调整，提升整体代码质量

+++ ydsz-common/ydsz-common-event/src/main/java/com/njydsz/common/event/processor/OutboxProcessor.java
@@ -406,8 +406,8 @@ public class OutboxProcessor {
    * @param retryCount 当前重试次数
    * @return 退避秒数
    */
-  private long calculateBackoff(int retryCount) {
-    int shift = Math.min(retryCount, MAX_SHIFT);
+  private long calculateBackoff(long retryCount) {
+    int shift = Math.min((int) retryCount, MAX_SHIFT);
     long backoff = properties.getBaseBackoffSeconds() * (1L << shift);
     return Math.min(backoff, properties.getMaxBackoffSeconds());
   }
